# legion/operator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a EDI server.",
        "title": "EDI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/v1/vcs": {
            "get": {
                "description": "Get list of VCS Credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VCSCredential"
                ],
                "summary": "Get list of VCS Credentials",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.VCSEntity"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a VCS Credential. Result is updated VCS Credential.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VCSCredential"
                ],
                "summary": "Update a VCS Credential",
                "parameters": [
                    {
                        "description": "Update a VCS Credential",
                        "name": "vcs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.VCSEntity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a VCS Credential. Result is created VCS Credential.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VCSCredential"
                ],
                "summary": "Create a VCS Credential",
                "parameters": [
                    {
                        "description": "Create a VCS Credential",
                        "name": "vcs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.VCSEntity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.VCSEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
        "/api/v1/vcs/{name}": {
            "get": {
                "description": "Get a VCS Credential by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VCSCredential"
                ],
                "summary": "Get a VCS Credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VCS Credential name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.VCSEntity"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a VCS Credential by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VCSCredential"
                ],
                "summary": "Delete a VCS Credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VCS Credential name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "routes.HTTPResult": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success of error message",
                    "type": "string"
                }
            }
        },
...
        "v1.VCSEntity": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "VCS name",
                    "type": "string"
                },
                "spec": {
                    "description": "VCS specification. It is the same as VCSCredential CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.VCSCredentialSpec"
                }
            }
        },
...
        "v1alpha1.VCSCredentialSpec": {
            "type": "object",
            "properties": {
                "credential": {
                    "description": "Creds for VCS. Is not required. In case of GIT should be base64-encoded private key.",
                    "type": "string"
                },
                "defaultReference": {
                    "description": "Default reference in VCS, e.g. branch, commit, tag and etc.",
                    "type": "string"
                },
                "publicKey": {
                    "description": "Public keys in base64 format for ssh know hosts. You can gather it using \"ssh-keyscan\"",
                    "type": "string"
                },
                "type": {
                    "description": "Type of VCS. Currently supports only git.\n+kubebuilder:validation:Enum=git",
                    "type": "string"
                },
                "uri": {
                    "description": "VCS uri.\nFor Git valid formats are:\n* git@github.com:legion-platform/legion.git",
                    "type": "string"
                }
            }
        }
    }
}